+++ base/deprecated.jl
@@ -1443,6 +1443,10 @@ end
 @deprecate digits(n, base, pad) digits(n, base = base, pad = pad)
 @deprecate digits(T, n, base, pad) digits(T, n, base = base, pad = pad)
 
+#27908
+@deprecate ndigits(n, base)      ndigits(n, base=base)
+@deprecate ndigits(n, base, pad) ndigits(n, base=base, pad=pad)
+
 @deprecate print_with_color(color, args...; kwargs...) printstyled(args...; kwargs..., color=color)
 
 @deprecate base(b, n)      string(n, base = b)